+++ changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improve variable dumps in REPL
 - Bump BoxLang version to 1.17.0
+- BLIDE-328 Fix CFC stack frames
 
 ## [1.1.1] - 2026-03-19